requirements.txt
colorama>=0.1.0
discord.py>=2.5.2
python-dotenv>=1.0.0
typing-extensions>=4.13.2
requests>=2.31.0
